Refactored order view into new component
